Fix Snake step parameter type from i64 to i32
The error "Cannot convert 3 to a BigInt" occurred because i64 in Rust
maps to BigInt in JavaScript/wasm-bindgen, but we were passing regular
JavaScript numbers from the game loop.

Changed the step(action: i64) parameter to i32, which maps to JavaScript's
number type. The action is then cast to i64 internally when calling the
Environment trait method.

Also added snake-gameplay.spec.ts test to verify game loop functionality.

This fixes the runtime error that occurred when clicking Start on the
Snake game.

+++ src/wasm.rs
@@ -120,9 +120,9 @@ impl WasmSnake {
     /// Step the environment with a single action
     /// action: 0 = up, 1 = down, 2 = left, 3 = right
     #[wasm_bindgen]
-    pub fn step(&mut self, action: i64) {
+    pub fn step(&mut self, action: i32) {
         use crate::env::Environment;
-        let _ = self.env.step(action);
+        let _ = self.env.step(action as i64);
     }
 
     /// Get observation for a specific agent
